feat: convert main to serverless api call
