+++ .gitignore
@@ -25,3 +25,4 @@ yt-dlp/
 
 /tmp/
 cache.json
+.vercel
