+++ api/hello.ts
@@ -1,0 +1,8 @@
+import type { VercelRequest, VercelResponse } from '@vercel/node'
+
+export default function handler(req: VercelRequest, res: VercelResponse) {
+    const { name = 'World' } = req.query
+    return res.json({
+        message: `Hello ${name}!`,
+    })
+}
