+++ tsconfig.json
@@ -4,10 +4,9 @@
         "outDir": "./build",
         "allowJs": true,
         "target": "es2022",
-        "module": "esnext",
         "moduleResolution": "node",
         "baseUrl": ".",
-        "rootDir": "src",
+        "rootDir": ".",
         "strictBindCallApply": true,
         "strictFunctionTypes": true,
         "strictNullChecks": true,
@@ -19,9 +18,5 @@
         "sourceMap": true,
         "skipLibCheck": true,
         "noUncheckedIndexedAccess": true,
-        "paths": {
-            "commander": ["./types/commander.d.ts"]
-        }
     },
-    "include": ["./src/**/*"]
 }
